Relocate ListDispatchConfigConverter to apriel.py; collapse decoder shim
Addresses /review-coarse @ 33971445:

* ``ConfigSectionConverter._consumed_hf_paths`` no longer special-cases primitive types via an
  isinstance switch. Each primitive that has dynamic claims (Nested, Dispatch, TypedDictContainer,
  ListDispatch) overrides ``ConfigConverter._consumed_hf_paths``; the aggregator collapses to a
  single union loop. The framework knows nothing about which primitives are "special".
* ``ListDispatchConfigConverter`` moves out of ``fast_llm/engine/checkpoint/external.py`` and into
  ``fast_llm/models/gpt/conversion/apriel.py`` next to its only consumer. The engine→layers import
  (``FixedBlockSequenceConfig`` / ``PatternBlockSequenceConfig``) goes away.
* The primitive's ``__init__`` now asserts ``set(layout_names) <= set(registry)`` so the
  load-bearing subset invariant is encoded in code rather than prose.
* ``AprielDecoderConverter`` is deleted; ``AprielBaseModelConverter`` carries the dispatch
  registry directly (``block_dispatcher_class``, named to avoid colliding with the parent's
  ``block_converter_class: type[ConfigSectionConverter]``) and inlines the per-block-index
  weight loop into its own ``get_converters``.
