+++ fast_llm/engine/checkpoint/external.py
@@ -14,7 +14,6 @@ from fast_llm.engine.checkpoint.config import CheckpointLoadMetadataConfig
 from fast_llm.engine.checkpoint.state_dict import StateDictCheckpointHandler
 from fast_llm.engine.multi_stage.config import CheckpointMetadata, FastLLMModelConfig
 from fast_llm.engine.multi_stage.fast_llm_model import FastLLMModel
-from fast_llm.layers.block.config import FixedBlockSequenceConfig, PatternBlockSequenceConfig
 from fast_llm.tensor import SafeTensorSlice
 from fast_llm.utils import Assert
 
@@ -117,6 +116,15 @@ class ConfigConverter(abc.ABC):
 
     @abc.abstractmethod
     def import_to(self, hf_dict: dict, fast_llm_out: dict) -> None: ...
+
+    def _consumed_hf_paths(self) -> frozenset[tuple[str, ...]]:
+        """HF paths this declaration claims for the section's coverage check.
+
+        Default: every non-empty entry in ``hf_paths``. Override when the claim set depends on
+        a sub-converter registry, a nested-path prefix, or a discriminator key (Nested, Dispatch,
+        TypedDictContainer, ListDispatch).
+        """
+        return frozenset(path for path in self.hf_paths if path)
 
 
 class RenameConfigConverter(ConfigConverter):
@@ -374,6 +382,21 @@ class NestedConfigConverter(ConfigConverter):
         sub_fast_llm = self._converter_class.import_config(sub_hf)
         _safe_set_nested_dict_value(fast_llm_out, self.fast_llm_paths[0], sub_fast_llm)
 
+    def _consumed_hf_paths(self) -> frozenset[tuple[str, ...]]:
+        sub_class = self._converter_class
+        if self._hf_path is None:
+            # Flat-merge: sub-converter shares the parent's HF namespace.
+            paths = set(sub_class._consumed_hf_paths())
+            if sub_class.hf_type_name is not None:
+                paths.add((self._hf_discriminator_key,))
+        else:
+            # Nested: prepend hf_path so the walker descends into the subdict and flags unknown keys.
+            prefix = self._hf_path
+            paths = {prefix + sub_path for sub_path in sub_class._consumed_hf_paths()}
+            if sub_class.hf_type_name is not None:
+                paths.add(prefix + (self._hf_discriminator_key,))
+        return frozenset(paths)
+
 
 class DispatchConfigConverter(ConfigConverter):
     """Polymorphic sub-config dispatch.
@@ -424,87 +447,22 @@ class DispatchConfigConverter(ConfigConverter):
         sub_fast_llm = converter_class.import_config(sub_hf)
         _safe_set_nested_dict_value(fast_llm_out, self.fast_llm_paths[0], sub_fast_llm)
 
-
-class ListDispatchConfigConverter(ConfigConverter):
-    """Polymorphic block-sequence dispatch driven by a positional HF list.
-
-    Models Apriel's hybrid-SSM shape: the HF config carries one entry per pipeline position in a
-    ``hybrid_block_layout: list[str]``, and the per-block HF keys live flat at the top level — each
-    registered block converter claims a disjoint subset (attention → flat top-level keys, mamba →
-    ``ssm_cfg.*``, gdn/kda → ``linear_attn_config.*``). On import, the layout list dispatches each
-    Fast-LLM block; on export, the block sequence (Fixed or Pattern) emits the union of each distinct
-    block's HF dict plus the layout and block count. Two registries are required because the format
-    distinguishes "which mixer classes appear at all" from "which mixer classes participate in the
-    config round-trip":
-
-    * ``registry``: mixer-config class → block ``ConfigSectionConverter``. Covers every block type
-      whose HF keys must be claimed by the coverage walker — including weight-side-only types whose
-      config-side layout name is absent (e.g. ``KimiDeltaAttentionConfig``).
-    * ``layout_names``: mixer-config class → layout discriminator string. Subset of ``registry`` keys
-      that participate in the config-side round-trip; classes absent here can't be config-exported
-      (no layout name to emit) or config-imported (no reverse-map entry).
-
-    Fast-LLM target is a :class:`FixedBlockSequenceConfig` when the layout has a single entry, a
-    :class:`PatternBlockSequenceConfig` otherwise. The same HF dict feeds every per-block import —
-    each block converter's declarations claim disjoint top-level keys, and shared top-level scalars
-    (e.g. ``rms_norm_eps``) cross-validate via the safe-set guard.
-    """
-
-    fast_llm_recurses: typing.ClassVar[bool] = True
-
-    def __init__(
-        self,
-        fast_llm_path: tuple[str, ...],
-        hf_layout_path: tuple[str, ...],
-        hf_num_blocks_path: tuple[str, ...],
-        registry: "dict[type[Config], type[ConfigSectionConverter]]",
-        layout_names: "dict[type[Config], str]",
-    ):
-        self.fast_llm_paths = (fast_llm_path,)
-        # Layout/num-blocks paths and per-block claims register through ``_consumed_hf_paths``.
-        self.hf_paths = ()
-        self._hf_layout_path = hf_layout_path
-        self._hf_num_blocks_path = hf_num_blocks_path
-        self._registry = registry
-        self._layout_names = layout_names
-        self._config_classes = {name: cls for cls, name in layout_names.items()}
-
-    def export_to(self, fast_llm_config: Config, hf_out: dict) -> None:
-        block_sequence = _get_attr_path(fast_llm_config, self.fast_llm_paths[0])
-        if type(block_sequence) is FixedBlockSequenceConfig:
-            distinct_blocks = [block_sequence.block]
-            pattern_blocks = [block_sequence.block]
-        elif type(block_sequence) is PatternBlockSequenceConfig:
-            distinct_blocks = list(block_sequence.blocks.values())
-            pattern_blocks = [block_sequence.blocks[name] for name in block_sequence.pattern]
+    def _consumed_hf_paths(self) -> frozenset[tuple[str, ...]]:
+        # Union of all registered sub-classes' claims (under the shared hf_path prefix when present).
+        # At runtime only one sub-class fires; the static union is a safe over-claim — we only need to
+        # not flag known keys, never to flag missing ones.
+        paths: set[tuple[str, ...]] = set()
+        if self.hf_paths:
+            prefix = self.hf_paths[0]
+            paths.add(prefix + (self._hf_discriminator_key,))
+            for sub_class in self._registry.values():
+                for sub_path in sub_class._consumed_hf_paths():
+                    paths.add(prefix + sub_path)
         else:
-            raise NotImplementedError(type(block_sequence).__name__)
-        for block_config in distinct_blocks:
-            converter_class = self._registry[type(block_config.mixer)]
-            sub_hf = converter_class.export_config(block_config)
-            for key, value in sub_hf.items():
-                _safe_set_nested_dict_value(hf_out, (key,), value)
-        layout = [self._layout_names[type(block_config.mixer)] for block_config in pattern_blocks]
-        _safe_set_nested_dict_value(hf_out, self._hf_layout_path, layout)
-        _safe_set_nested_dict_value(hf_out, self._hf_num_blocks_path, block_sequence.num_blocks)
-
-    def import_to(self, hf_dict: dict, fast_llm_out: dict) -> None:
-        layout = get_nested_dict_value(hf_dict, self._hf_layout_path)
-        num_blocks = get_nested_dict_value(hf_dict, self._hf_num_blocks_path)
-        if len(layout) == 1:
-            converter_class = self._registry[self._config_classes[layout[0]]]
-            sub_fast_llm = {"block": converter_class.import_config(hf_dict), "num_blocks": num_blocks}
-        else:
-            sub_fast_llm = {
-                "type": "pattern",
-                "blocks": {
-                    layout_name: self._registry[self._config_classes[layout_name]].import_config(hf_dict)
-                    for layout_name in set(layout)
-                },
-                "pattern": layout,
-                "num_blocks": num_blocks,
-            }
-        _safe_set_nested_dict_value(fast_llm_out, self.fast_llm_paths[0], sub_fast_llm)
+            paths.add((self._hf_discriminator_key,))
+            for sub_class in self._registry.values():
+                paths |= sub_class._consumed_hf_paths()
+        return frozenset(paths)
 
 
 class TypedDictContainerConfigConverter(ConfigConverter):
@@ -570,6 +528,11 @@ class TypedDictContainerConfigConverter(ConfigConverter):
             sub_fast_llm = converter_class.import_config(sub_hf)
             out[name] = sub_fast_llm
         _safe_set_nested_dict_value(fast_llm_out, self.fast_llm_paths[0], out)
+
+    def _consumed_hf_paths(self) -> frozenset[tuple[str, ...]]:
+        # Blanket prefix: per-entry sub-dicts are user-named pattern keys; we can't statically
+        # enumerate which entries appear or what keys those entries claim.
+        return frozenset({self.hf_paths[0]})
 
 
 # ============================================================
@@ -652,56 +615,14 @@ class ConfigSectionConverter(abc.ABC):
         walker treats every entry as a *recursive prefix* — once an input path matches any prefix,
         descent into deeper sub-dicts stops.
 
-        Nested sub-converters (``NestedConfigConverter``/``DispatchConfigConverter`` with ``hf_path``
-        set) expand their sub-converter's claims under the nested prefix instead of contributing the
-        bare prefix, so the walker descends into the subdict and flags unknown keys inside it (e.g.
-        ``head.normalization.surprise_field``).
-
-        :class:`TypedDictContainerConfigConverter` keeps a blanket prefix because its per-entry sub-dicts
-        are user-named (pattern keys); we can't statically enumerate which entries will appear or what
-        keys those entries should claim.
+        Each declaration advertises its claims via :meth:`ConfigConverter._consumed_hf_paths`
+        (default: ``hf_paths``; overridden by primitives whose claims depend on a sub-converter
+        registry, a nested-path prefix, or a discriminator key). The aggregate is cached per
+        section class.
         """
         paths: set[tuple[str, ...]] = set()
         for declaration in cls._create_config_converters().values():
-            if isinstance(declaration, NestedConfigConverter):
-                sub_class = declaration._converter_class
-                if declaration._hf_path is None:
-                    # Flat-merge: sub-converter shares the parent's HF namespace.
-                    paths |= sub_class._consumed_hf_paths()
-                    if sub_class.hf_type_name is not None:
-                        paths.add((declaration._hf_discriminator_key,))
-                else:
-                    # Nested: prepend hf_path to each sub-claim so the walker recurses into the subdict.
-                    prefix = declaration._hf_path
-                    for sub_path in sub_class._consumed_hf_paths():
-                        paths.add(prefix + sub_path)
-                    if sub_class.hf_type_name is not None:
-                        paths.add(prefix + (declaration._hf_discriminator_key,))
-            elif isinstance(declaration, DispatchConfigConverter):
-                if declaration.hf_paths:
-                    # Nested dispatch: union of all registered sub-classes' claims under the shared
-                    # hf_path prefix. At runtime only one sub-class fires; the static union is a safe
-                    # over-claim (we only need to *not* flag known keys, never to flag missing ones).
-                    prefix = declaration.hf_paths[0]
-                    paths.add(prefix + (declaration._hf_discriminator_key,))
-                    for sub_class in declaration._registry.values():
-                        for sub_path in sub_class._consumed_hf_paths():
-                            paths.add(prefix + sub_path)
-                else:
-                    paths.add((declaration._hf_discriminator_key,))
-                    for sub_class in declaration._registry.values():
-                        paths |= sub_class._consumed_hf_paths()
-            elif isinstance(declaration, TypedDictContainerConfigConverter):
-                paths.add(declaration.hf_paths[0])
-            elif isinstance(declaration, ListDispatchConfigConverter):
-                paths.add(declaration._hf_layout_path)
-                paths.add(declaration._hf_num_blocks_path)
-                for sub_class in declaration._registry.values():
-                    paths |= sub_class._consumed_hf_paths()
-            else:
-                for path in declaration.hf_paths:
-                    if path:
-                        paths.add(path)
+            paths |= declaration._consumed_hf_paths()
         return frozenset(paths)
 
     @classmethod
